+++ pricing/src/epilot/models/operations/dollar_checkoutcart.py
@@ -12,7 +12,7 @@ from typing import Optional
 
 @dataclasses.dataclass
 class DollarCheckoutCartRequest:
-    checkout_cart_input: shared_checkoutcart.CheckoutCartInput = dataclasses.field(metadata={'request': { 'media_type': 'application/json' }})
+    checkout_cart: shared_checkoutcart.CheckoutCart = dataclasses.field(metadata={'request': { 'media_type': 'application/json' }})
     x_ivy_org_id: str = dataclasses.field(metadata={'header': { 'field_name': 'X-Ivy-Org-ID', 'style': 'simple', 'explode': False }})
     r"""The target Organization Id represented by the caller"""